Make sentinels avoid bombs

+++ src/com/megahard/gravity/engine/GameContext.java
@@ -71,8 +71,8 @@ public interface GameContext {
 	public abstract void fadeScreen(Color colorStart, Color colorEnd,
 			int duration);
 
-	public abstract <T> List<T> findObjects(Class<T> type, int x, int y, int w,
-			int h, boolean inclusive);
+	public abstract <T> List<T> findObjects(Class<T> type, double x, double y, double w,
+			double h, boolean inclusive);
 
 	public abstract <T> List<T> findObjects(Class<T> type);
